# lib/api-zod/src/generated/types/emergencyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Scheduler 2 API specification
 * OpenAPI spec version: 0.2.0
 */
import type { Job } from "./job";

export interface EmergencyResponse {
  emergencyJob: Job;
  /** Number of bookings bumped */
  bumpedCount: number;
  bumpedJobIds: number[];
}