# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Scheduler 2 API specification
 * OpenAPI spec version: 0.2.0
 */
import type { Worker } from "./worker";

export interface Invoice {
  invoiceNumber: string;
  jobId: number;
  jobTitle: string;
  clientName: string;
  clientPhone?: string | null;
  clientEmail?: string | null;
  address: string;
  tradeType: string;
  estimatedHours: number;
  price: number;
  /** 10% GST (Australian) */
  gst: number;
  totalWithGst: number;
  scheduledDate?: string | null;
  completedDate?: string | null;
  status: string;
  issuedAt: string;
  notes?: string | null;
  assignedWorkers: Worker[];
}