# lib/api-zod/src/generated/types/jobType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Scheduler 2 API specification
 * OpenAPI spec version: 0.2.0
 */

export type JobType = (typeof JobType)[keyof typeof JobType];

export const JobType = {
  quote: "quote",
  booking: "booking",
} as const;