# trade-scheduler/backend/src/api-zod/generated/types/convertToBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Scheduler 2 API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ConvertToBookingRequest {
  estimatedHours: number;
  numTradies?: number;
  callUpTimeHours?: number | null;
  scheduledDate?: string | null;
  assignedWorkerIds?: number[];
  notes?: string | null;
}